+++ src/evalBlock.ts
@@ -16,7 +16,7 @@ const evalFuncTemplate = `
 
   // Note this feature is currently unstable. The final structure of forkMap
   // is likely to change before stabilizing.
-  bml.forkMap = __context.renderer.choiceResultMap;
+  bml.forkMap = __context.renderer.executedForkMap;
 
   ***USER DEFS BINDING SLOT***